AtkCounter & HpBar Bug Fix
Added possibility of attacks destroying other attacks
Fixed the bug where the hp bar would go to zero with any damage byu changing the health to float

+++ Assets/Scripts/Health.cs
@@ -6,14 +6,15 @@ using UnityEngine.UI;
 public class Health : MonoBehaviour
 {
     [Header("Values")]
-    public int health;
+    public float health;
     public int maxHealth;
     public Image healthBar;
 
     private void Start(){
         health = maxHealth;
     }
-    public void Dmg(int dmg){
+    public void Dmg(int dmg)
+    {
         health -= dmg;
         healthBar.fillAmount = health / 100;
         Debug.Log($"{health},{gameObject}");
